Typescript migration - it compiles
