+++ app/package.json
@@ -4,6 +4,10 @@
   "private": true,
   "dependencies": {
     "@material-ui/core": "^3.9.2",
+    "@types/node": "^12.0.2",
+    "@types/react": "^16.8.18",
+    "@types/react-dom": "^16.8.4",
+    "@types/react-router-dom": "^4.3.3",
     "axios": "^0.18.0",
     "cryptocompare": "^1.0.0",
     "drizzle": "https://github.com/simondlr/drizzle.git#ee8ff32",
@@ -14,10 +18,11 @@
     "react-dom": "^16.7.0",
     "react-id-swiper": "^2.1.2",
     "react-router-dom": "^5.0.0",
-    "react-scripts": "^2.1.8",
+    "react-scripts": "^3.0.1",
     "rimble-ui": "^0.6.1",
     "styled-components": "^4.2.0",
-    "swiper": "^4.5.0"
+    "swiper": "^4.5.0",
+    "typescript": "^3.4.5"
   },
   "scripts": {
     "start": "react-scripts start",
