+++ app/src/components/AboutArtSection.tsx
