+++ app/src/components/ActionSection.tsx
@@ -7,9 +7,18 @@ import ContractForm from "./ContractForm";
 import DepositWeiForm from "./DepositWeiForm";
 import ContractData from "./ContractData";
 
-class ActionSection extends Component {
-  constructor(props, context) {
-    super();
+declare global {
+  interface Window { ethereum: any; }
+}
+
+window.ethereum = window.ethereum || {}
+
+class ActionSection extends Component<{ contracts: any }, {}>  {
+  utils: any
+  contracts: any
+  state: any
+  constructor(props: any, context: any) {
+    super(props, context)
     this.utils = context.drizzle.web3.utils;
     this.contracts = context.drizzle.contracts;
     this.state = {
@@ -21,11 +30,11 @@ class ActionSection extends Component {
     };
   }
 
-  getTotalCollected(props) {
+  getTotalCollected(props: any) {
     return new this.utils.BN(props.contracts['ArtSteward']['totalCollected'][this.state.totalCollectedKey].value);
   }
 
-  async updateCombineCollected(props) {
+  async updateCombineCollected(props: any) {
     const patronageOwed = this.getPatronageOwed(props);
     const totalCollected = this.getTotalCollected(props);
     const combinedCollected = this.utils.fromWei(totalCollected.add(patronageOwed), 'ether').toString();
@@ -35,26 +44,26 @@ class ActionSection extends Component {
     });
   }
 
-  getPatronageOwed(props) {
+  getPatronageOwed(props: any) {
     return new this.utils.BN(props.contracts['ArtSteward']['patronageOwed'][this.state.patronageOwedKey].value);
   }
 
-  async componentWillReceiveProps(nextProps) {
-    if(this.props.contracts['ArtSteward']['price']['0x0'] !== nextProps.contracts['ArtSteward']['price']['0x0']) {
+  async componentWillReceiveProps(nextProps: any) {
+    if (this.props.contracts['ArtSteward']['price']['0x0'] !== nextProps.contracts['ArtSteward']['price']['0x0']) {
       if (nextProps.contracts['ArtSteward']['price']['0x0'].value === '0') {
         this.setState({
           foreclosureTime: "N/A"
         });
       } else {
-        const foreclosureTime = moment(parseInt(await this.contracts.ArtSteward.methods.foreclosureTime().call())*1000).toString();
-        this.setState({foreclosureTime});
+        const foreclosureTime = moment(parseInt(await this.contracts.ArtSteward.methods.foreclosureTime().call()) * 1000).toString();
+        this.setState({ foreclosureTime });
       }
     }
 
     if (this.state.patronageOwedKey in this.props.contracts['ArtSteward']['patronageOwed']
-    && this.state.patronageOwedKey in nextProps.contracts['ArtSteward']['patronageOwed']
-    && this.state.totalCollectedKey in this.props.contracts['ArtSteward']['totalCollected']
-    && this.state.totalCollectedKey in nextProps.contracts['ArtSteward']['totalCollected']) {
+      && this.state.patronageOwedKey in nextProps.contracts['ArtSteward']['patronageOwed']
+      && this.state.totalCollectedKey in this.props.contracts['ArtSteward']['totalCollected']
+      && this.state.totalCollectedKey in nextProps.contracts['ArtSteward']['totalCollected']) {
       if (!this.getPatronageOwed(this.props).eq(this.getPatronageOwed(nextProps)) || this.state.combinedCollected === -1) {
         this.updateCombineCollected(nextProps);
       }
@@ -63,47 +72,40 @@ class ActionSection extends Component {
 
   render() {
     return (
-    <div className="section">
-      <h2>Current Patron Details:</h2>
-        <p>Address: <ContractData contract="ERC721Full" method="ownerOf" methodArgs={[42]}/></p>
+      <div className="section">
+        <h2>Current Patron Details:</h2>
+        <p>Address: <ContractData contract="ERC721Full" method="ownerOf" methodArgs={[42]} /></p>
         <p>Available Deposit: <ContractData contract="ArtSteward" method="depositAbleToWithdraw" toEth /> ETH</p>
         <p>Foreclosure Time: {this.state.foreclosureTime}</p>
         <p>The current deposit will cover the patronage until the time above. At this time, the smart contract steward takes ownership of the artwork and sets its price back to zero.</p>
         <p>Once it crosses this time period, the patron can't top up their deposit anymore and is effectively foreclosed.</p>
-      <h2>Actions:</h2>
+        <h2>Actions:</h2>
         {window.ethereum !== undefined ? (
           <Fragment>
-          <ContractForm buttonText="Change Price" contract="ArtSteward" method="changePrice" labels={["New Price"]}/>
-          <DepositWeiForm contract="ArtSteward" method="depositWei" valueLabel="Deposit" sendArgs={{}}/>
-          <ContractForm buttonText="Withdraw Deposit" contract="ArtSteward" method="withdrawDeposit" labels={["Deposit in ETH"]} toEth />
-          <ContractForm buttonText="Withdraw Whole Deposit And Foreclose" contract="ArtSteward" method="exit" />
+            <ContractForm buttonText="Change Price" contract="ArtSteward" method="changePrice" labels={["New Price"]} />
+            <DepositWeiForm contract="ArtSteward" method="depositWei" valueLabel="Deposit" sendArgs={{}} />
+            <ContractForm buttonText="Withdraw Deposit" contract="ArtSteward" method="withdrawDeposit" labels={["Deposit in ETH"]} toEth />
+            <ContractForm buttonText="Withdraw Whole Deposit And Foreclose" contract="ArtSteward" method="exit" />
           </Fragment>
         ) : (
-          <Fragment>
-          In order to interact with the artwork you need to have a web3/Ethereum-enabled browser. Please download
+            <Fragment>
+              In order to interact with the artwork you need to have a web3/Ethereum-enabled browser. Please download
           the <a href="https://metamask.io">MetaMask Chrome extension</a> or open in an Ethereum mobile browser.
           </Fragment>
-        )}
+          )}
 
-      <h2>Other Artwork Stats:</h2>
+        <h2>Other Artwork Stats:</h2>
         <p>Total Patronage Collected: {this.state.combinedCollected} ETH</p>
-    </div>
+      </div>
     )
   }
 }
-
-ActionSection.contextTypes = {
-  drizzle: PropTypes.object,
-};
-
-ActionSection.propTypes = {
-};
 
 /*
  * Export connected component.
  */
 
-const mapStateToProps = state => {
+const mapStateToProps = (state: any) => {
   return {
     contracts: state.contracts,
   };
