+++ app/src/components/AnimalAdoptionApp/AnimalCardContainer.tsx
@@ -1,0 +1,270 @@
+import React, { Component, ReactElement } from 'react'
+import { render } from 'react-dom'
+import PropTypes from 'prop-types'
+import FooterBarNav from './FooterBarNav'
+import AnimalCard from './AnimalCard'
+import ClosingInformationCard from './ClosingInformationCard'
+import OpeningInformationCard from './OpeningInformationCard'
+import CollectionPage from './CollectionPage/CollectionPage'
+
+import Swiper from 'react-id-swiper/lib/ReactIdSwiper.full';
+import './../../App.css'
+import { ReactIdSwiperProps } from 'react-id-swiper/lib/types';
+// import { Pagination, Navigation } from 'swiper/dist/js/swiper.esm'
+
+class AnimalCardContainer extends Component<{}, { cards: any[], showStore: boolean, showCollection: boolean }> {
+
+  constructor(props: any) {
+    super(props);
+    this.state = {
+      showStore: true,
+      showCollection: false,
+      cards: [
+        {
+          id: 0,
+          name: "Vitalik", // let user set this on first buy
+          age: 12,
+          lineage: "african",
+          weight: 900,
+          is_female: false,
+          img_source_url: "myurl",
+
+          // Values we will be pulling from chain
+          // hardcoding them for now...
+          price: 100,
+          current_owner: 0x348341035837599834,
+          time_held: 12, //must define if this is hours days etc...
+          highest_price: 12314,
+          revenue_generated: 9
+        },
+        {
+          id: 1,
+          name: "Andrew", // let user set this on first buy
+          age: 12,
+          lineage: "african",
+          weight: 900,
+          is_female: false,
+          img_source_url: "myurl",
+
+          // Values we will be pulling from chain
+          // hardcoding them for now...
+          price: 100,
+          current_owner: 0x348341035837599834,
+          time_held: 12, //must define if this is hours days etc...
+          highest_price: 12314,
+          revenue_generated: 9
+        },
+        {
+          id: 2,
+          name: "Simon", // let user set this on first buy
+          age: 12,
+          lineage: "african",
+          weight: 900,
+          is_female: true,
+          img_source_url: "myurl",
+
+          // Values we will be pulling from chain
+          // hardcoding them for now...
+          price: 100,
+          current_owner: 0x348341035837599834,
+          time_held: 12, //must define if this is hours days etc...
+          highest_price: 12314,
+          revenue_generated: 9
+        },
+        {
+          id: 3,
+          name: "Griff", // let user set this on first buy
+          age: 12,
+          lineage: "african",
+          weight: 900,
+          is_female: false,
+          img_source_url: "myurl",
+
+          // Values we will be pulling from chain
+          // hardcoding them for now...
+          price: 100,
+          current_owner: 0x348341035837599834,
+          time_held: 12, //must define if this is hours days etc...
+          highest_price: 12314,
+          revenue_generated: 9
+        },
+        {
+          id: 4,
+          name: "Anastasia", // let user set this on first buy
+          age: 12,
+          lineage: "african",
+          weight: 900,
+          is_female: true,
+          img_source_url: "myurl",
+
+          // Values we will be pulling from chain
+          // hardcoding them for now...
+          price: 100,
+          current_owner: 0x348341035837599834,
+          time_held: 12, //must define if this is hours days etc...
+          highest_price: 12314,
+          revenue_generated: 9
+        },
+        {
+          id: 5,
+          name: "Justin", // let user set this on first buy
+          age: 12,
+          lineage: "african",
+          weight: 900,
+          is_female: false,
+          img_source_url: "myurl",
+
+          // Values we will be pulling from chain
+          // hardcoding them for now...
+          price: 100,
+          current_owner: 0x348341035837599834,
+          time_held: 12, //must define if this is hours days etc...
+          highest_price: 12314,
+          revenue_generated: 9
+        },
+        {
+          id: 6,
+          name: "Michael", // let user set this on first buy
+          age: 12,
+          lineage: "african",
+          weight: 900,
+          is_female: false,
+          img_source_url: "myurl",
+
+          // Values we will be pulling from chain
+          // hardcoding them for now...
+          price: 100,
+          current_owner: 0x348341035837599834,
+          time_held: 12, //must define if this is hours days etc...
+          highest_price: 12314,
+          revenue_generated: 9
+        },
+        {
+          id: 7,
+          name: "Gustav", // let user set this on first buy
+          age: 12,
+          lineage: "african",
+          weight: 900,
+          is_female: true,
+          img_source_url: "myurl",
+
+          // Values we will be pulling from chain
+          // hardcoding them for now...
+          price: 100,
+          current_owner: 0x348341035837599834,
+          time_held: 12, //must define if this is hours days etc...
+          highest_price: 12314,
+          revenue_generated: 9
+        },
+        {
+          id: 8,
+          name: "George", // let user set this on first buy
+          age: 12,
+          lineage: "african",
+          weight: 900,
+          is_female: false,
+          img_source_url: "myurl",
+
+          // Values we will be pulling from chain
+          // hardcoding them for now...
+          price: 100,
+          current_owner: 0x348341035837599834,
+          time_held: 12, //must define if this is hours days etc...
+          highest_price: 12314,
+          revenue_generated: 9
+        },
+        {
+          id: 9,
+          name: "Sean", // let user set this on first buy
+          age: 12,
+          lineage: "african",
+          weight: 900,
+          is_female: false,
+          img_source_url: "myurl",
+
+          // Values we will be pulling from chain
+          // hardcoding them for now...
+          price: 100,
+          current_owner: 0x348341035837599834,
+          time_held: 12, //must define if this is hours days etc...
+          highest_price: 12314,
+          revenue_generated: 9
+        },
+
+      ]
+    }
+  }
+
+  renderCards() {
+    const cardStyle = {
+      backgroundColor: "#2a3132",
+      height: '80vh',
+      marginTop: '20px'
+    }
+    const startingCard =
+      <div key={'initialkey'}>
+        <OpeningInformationCard />
+      </div> as ReactElement
+    const finalCard =
+      <div key={'finalkey'}>
+        <ClosingInformationCard />
+      </div> as ReactElement
+    const cards = this.state.cards.map((specificCard, index) => {
+      return (
+        <div key={'mykey' + index}>
+          <AnimalCard Rhino={specificCard} />
+        </div> as ReactElement
+      )
+    })
+
+    return [startingCard, ...cards, finalCard]
+  }
+
+  clickShowCollection = () => {
+    this.setState({ showCollection: true });
+  }
+  clickHideCollection = () => {
+    this.setState({ showCollection: false });
+  }
+
+  render() {
+
+    const params = {
+      pagination: {
+        el: '.swiper-pagination',
+        clickable: true,
+        dynamicBullets: true
+      }
+    } as ReactIdSwiperProps
+
+    const containerViewPort = {
+      backgroundImage: 'linear-gradient(120deg, #a1c4fd 0%, #c2e9fb 100%)',
+      width: '100%',
+      height: '100vh',
+      position: 'absolute',
+      top: 0,
+      left: 0
+    } as React.CSSProperties
+
+    return (
+
+      <div style={containerViewPort}>
+        {
+          this.state.showCollection ?
+
+            <CollectionPage />
+            :
+            <Swiper {...params}>
+              {this.renderCards()}
+            </Swiper>
+        }
+        <FooterBarNav
+          clickShowCollection={this.clickShowCollection}
+          clickHideCollection={this.clickHideCollection}
+        />
+      </div>
+    );
+  }
+}
+
+export default AnimalCardContainer;
