+++ app/src/components/ArtistSection.tsx
