+++ app/src/components/BuyingArtSection.tsx
