+++ app/src/components/DepositWeiForm.tsx
@@ -3,21 +3,19 @@ import React, { Component, Fragment } from "react";
 import PropTypes from "prop-types";
 import Input from '@material-ui/core/Input';
 import InputAdornment from '@material-ui/core/InputAdornment';
-// import Button from '@material-ui/core/Button';
-import { Card, Button, Modal, TextButton, Box, Flex, OutlineButton, Text, Heading, Icon } from 'rimble-ui'
+import Button from '@material-ui/core/Button';
 
 /*
 Edited from drizzle react components, ContractFrom.
 Overkill. Needs to be refactored to smaller scope.
 */
 
-/*
-Update: Edited again from This Artwork is always on sale. For Hackathon. 
-Needs to be refactored to smaller smaller scope. Down the rabbit hole.
- */
-
-class BuyForm extends Component {
-  constructor(props, context) {
+class BuyForm extends Component<{ contract: any, method: any, sendArgs: any, valueLabel: any, labels: any[] }> {
+  contracts: any
+  utils: any
+  inputs: any[]
+  state: any
+  constructor(props: any, context: any) {
     super(props);
 
     this.handleInputChange = this.handleInputChange.bind(this);
@@ -30,7 +28,7 @@ class BuyForm extends Component {
     const abi = this.contracts[this.props.contract].abi;
 
     this.inputs = [];
-    var initialState = {};
+    var initialState: any = {};
 
     // Iterate over abi for correct function.
     for (var i = 0; i < abi.length; i++) {
@@ -48,7 +46,7 @@ class BuyForm extends Component {
     this.state = initialState;
   }
 
-  handleSubmit(event) {
+  handleSubmit(event: any) {
     event.preventDefault();
     let args = this.props.sendArgs;
     const convertedInputs = this.inputs.map((input, index) => {
@@ -60,11 +58,9 @@ class BuyForm extends Component {
       return this.state[input.name];
     });
 
-    // todo: if foreclosed, price should default to zero.
     if (this.state.value) {
-      console.log(this.props.contracts[this.props.contract]['price']['0x0'].value);
-      const artworkPrice = new this.utils.BN(this.props.contracts[this.props.contract]['price']['0x0'].value);
-      args.value = new this.utils.BN(this.utils.toWei(this.state.value, 'ether')).add(artworkPrice);
+      // const artworkPrice = new this.utils.BN(this.props.contracts[this.props.contract]['price']['0x0'].value);
+      args.value = new this.utils.BN(this.utils.toWei(this.state.value, 'ether'));
     }
     if (args) {
       return this.contracts[this.props.contract].methods[
@@ -77,11 +73,11 @@ class BuyForm extends Component {
     ].cacheSend(...convertedInputs);
   }
 
-  handleInputChange(event) {
+  handleInputChange(event: any) {
     this.setState({ [event.target.name]: event.target.value });
   }
 
-  translateType(type) {
+  translateType(type: any) {
     switch (true) {
       case /^uint/.test(type):
         return "number";
@@ -105,76 +101,51 @@ class BuyForm extends Component {
             : input.name;
           // check if input type is struct and if so loop out struct fields as well
           return (
-              <center>
-
-            <Input              
-              style={{width: '100%'}}
+            <Input
               key={input.name}
               type={inputType}
               name={input.name}
               value={this.state[input.name]}
               placeholder={inputLabel}
               onChange={this.handleInputChange}
-              startAdornment={<InputAdornment position="start">ETH</InputAdornment>} 
-              />
-              </center>
+              startAdornment={<InputAdornment position="start">ETH</InputAdornment>}
+            />
           );
         })}
         {valueLabel &&
           <Fragment>
-          <br />
-          <center>
-
-          <Input 
-          style={{width: '100%'}}
-          key={valueLabel} 
-          type='number' 
-          name='value' 
-          value={this.state[valueLabel]} 
-          placeholder={valueLabel} 
-          onChange={this.handleInputChange} 
-          startAdornment={<InputAdornment position="start">ETH</InputAdornment>} />
-          </center>
-          <br />
-          <br />
+            <Input
+              key={valueLabel}
+              type='number'
+              name='value'
+              value={this.state[valueLabel]}
+              placeholder={valueLabel}
+              onChange={this.handleInputChange}
+              startAdornment={<InputAdornment position="start">ETH</InputAdornment>} />
           </Fragment>
+
         }
-         <Flex px={4} py={3} borderTop={1} borderColor={'#E8E8E8'} justifyContent={'flex-end'}>
-                            <OutlineButton>Cancel</OutlineButton>
         <Button
-          ml ={3}
           variant="contained"
           key="submit"
           className="pure-button"
           type="button"
           onClick={this.handleSubmit}
-          fullWidth
         >
-          Adopt Rhino
+          Top Up Deposit
         </Button>
-        </Flex>
+        <br />
+        <br />
       </form>
     );
   }
 }
 
-BuyForm.contextTypes = {
-  drizzle: PropTypes.object,
-};
-
-// todo: add value label
-BuyForm.propTypes = {
-  contract: PropTypes.string.isRequired,
-  method: PropTypes.string.isRequired,
-  sendArgs: PropTypes.object,
-  labels: PropTypes.arrayOf(PropTypes.string),
-};
-
 /*
  * Export connected component.
  */
 
-const mapStateToProps = state => {
+const mapStateToProps = (state: any) => {
   return {
     contracts: state.contracts,
   };
