+++ app/src/components/IntroSection.tsx
@@ -6,28 +6,21 @@ import BuyArtworkSection from "./BuyArtworkSection";
 import ArtAndPriceSection from "./ArtAndPriceSection";
 
 class IntroSection extends Component {
-    render() {
-      return (
-        <div className="section">
+  render() {
+    return (
+      <div className="section">
         <ArtAndPriceSection />
         <BuyArtworkSection />
       </div>
-      )
-    }
+    )
+  }
 }
-
-IntroSection.contextTypes = {
-  drizzle: PropTypes.object,
-};
-
-IntroSection.propTypes = {
-};
 
 /*
  * Export connected component.
  */
 
-const mapStateToProps = state => {
+const mapStateToProps = (state: any) => {
   return {
     contracts: state.contracts,
     drizzleStatus: state.drizzleStatus,
