+++ app/src/react-app-env.d.ts
@@ -1,0 +1,7 @@
+/// <reference types="react-scripts" />
+
+declare module 'drizzle';
+declare module 'drizzle-react';
+declare module 'drizzle-react-components';
+declare module 'rimble-ui';
+
